[pickers] Improve useSplitFieldInternalAndForwardedProps and make it public
