+++ packages/x-date-pickers/src/DateField/useDateField.ts
@@ -5,7 +5,7 @@ import {
 import { useField } from '../internals/hooks/useField';
 import { UseDateFieldProps } from './DateField.types';
 import { validateDate } from '../internals/utils/validation/validateDate';
-import { splitFieldInternalAndForwardedProps } from '../internals/utils/fields';
+import { useSplitFieldInternalAndForwardedProps } from '../hooks';
 import { FieldSection, PickerValidDate } from '../models';
 import { useDefaultizedDateField } from '../internals/hooks/defaultizedFieldProps';
 
@@ -22,10 +22,7 @@ export const useDateField = <
     TAllProps
   >(inProps);
 
-  const { forwardedProps, internalProps } = splitFieldInternalAndForwardedProps<
-    typeof props,
-    keyof UseDateFieldProps<TDate, TEnableAccessibleFieldDOMStructure>
-  >(props, 'date');
+  const { forwardedProps, internalProps } = useSplitFieldInternalAndForwardedProps(props, 'date');
 
   return useField<
     TDate | null,
