+++ packages/x-date-pickers/src/DateTimeField/useDateTimeField.ts
@@ -5,7 +5,7 @@ import {
 import { useField } from '../internals/hooks/useField';
 import { UseDateTimeFieldProps } from './DateTimeField.types';
 import { validateDateTime } from '../internals/utils/validation/validateDateTime';
-import { splitFieldInternalAndForwardedProps } from '../internals/utils/fields';
+import { useSplitFieldInternalAndForwardedProps } from '../hooks';
 import { FieldSection, PickerValidDate } from '../models';
 import { useDefaultizedDateTimeField } from '../internals/hooks/defaultizedFieldProps';
 
@@ -22,10 +22,10 @@ export const useDateTimeField = <
     TAllProps
   >(inProps);
 
-  const { forwardedProps, internalProps } = splitFieldInternalAndForwardedProps<
-    typeof props,
-    keyof UseDateTimeFieldProps<any, TEnableAccessibleFieldDOMStructure>
-  >(props, 'date-time');
+  const { forwardedProps, internalProps } = useSplitFieldInternalAndForwardedProps(
+    props,
+    'date-time',
+  );
 
   return useField<
     TDate | null,
