+++ packages/x-date-pickers/src/hooks/index.tsx
@@ -7,3 +7,5 @@ export type {
 } from './useClearableField';
 
 export { usePickersTranslations } from './usePickersTranslations';
+
+export { useSplitFieldInternalAndForwardedProps } from './useSplitFieldInternalAndForwardedProps';
