+++ scripts/x-date-pickers-pro.exports.json
@@ -418,6 +418,7 @@
   { "name": "UseSingleInputDateRangeFieldProps", "kind": "Interface" },
   { "name": "UseSingleInputDateTimeRangeFieldProps", "kind": "Interface" },
   { "name": "UseSingleInputTimeRangeFieldProps", "kind": "Interface" },
+  { "name": "useSplitFieldInternalAndForwardedProps", "kind": "Variable" },
   { "name": "UseTimeFieldComponentProps", "kind": "TypeAlias" },
   { "name": "UseTimeFieldProps", "kind": "Interface" },
   { "name": "YearCalendar", "kind": "Variable" },
