+++ scripts/x-date-pickers.exports.json
@@ -300,6 +300,7 @@
   { "name": "UseDateTimeFieldProps", "kind": "Interface" },
   { "name": "usePickerLayout", "kind": "ExportAssignment" },
   { "name": "usePickersTranslations", "kind": "Variable" },
+  { "name": "useSplitFieldInternalAndForwardedProps", "kind": "Variable" },
   { "name": "UseTimeFieldComponentProps", "kind": "TypeAlias" },
   { "name": "UseTimeFieldProps", "kind": "Interface" },
   { "name": "YearCalendar", "kind": "Variable" },
